refactor(frontend/register): add register module
