+++ packages/frontend/src/components/containers/index.ts
@@ -1,7 +1,6 @@
-import RegisterForm from './RegisterForm';
 import PayPal from './PayPal';
 import CartContent from './CartContent';
 import UserForms from './UserForms';
 import UserProducts from './UserProducts';
 
-export { RegisterForm, PayPal, CartContent, UserForms, UserProducts };
+export { PayPal, CartContent, UserForms, UserProducts };
