+++ packages/frontend/src/components/modules/Register/Register.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Redirect } from 'react-router';
 import Cookies from 'universal-cookie';
-import { useAuth } from '../../hooks';
-import { Loader } from '../elements';
+import { useAuth } from '../../../hooks';
+import { Loader } from '../../elements';
 
-import { RegisterForm } from '../containers';
-import { Nav } from '../elements';
+import RegisterForm from './RegisterForm';
+import { Nav } from '../../elements';
 
 const Register = () => {
   const cookie = new Cookies();
