+++ packages/frontend/src/components/pages/index.ts
@@ -1,5 +1,4 @@
 import Main from './Main';
-import Register from './Register';
 import Home from './Home';
 import Payment from './Payment';
 import Cart from './Cart';
@@ -7,4 +6,4 @@ import User from './User';
 import UserProducts from './UserProducts';
 import UserCart from './UserCart';
 
-export { Main, Register, Home, Payment, Cart, User, UserProducts, UserCart };
+export { Main, Home, Payment, Cart, User, UserProducts, UserCart };
